+++ client/src/features/auth/components/AuthOutlet.tsx
