+++ client/src/features/auth/components/AuthSidePanel.tsx
