+++ client/src/features/auth/components/register/SignUpCard.tsx
@@ -1,36 +1,31 @@
 import React, { MouseEvent, useState } from "react";
-import { Link } from "react-router-dom";
-import { SubmitHandler, useController, useForm } from "react-hook-form";
-import { ApolloError } from "@apollo/client";
-import { GoogleIcon, FacebookIcon } from "@lib/components/Auth/CustomIcons";
-import { ForgotPassword } from "./ForgotPassword";
-import { useLoginMutation } from "@pages/auth/hooks/useLoginMutation";
-import {
-  LoginSchema,
-  useLoginResolver,
-} from "@pages/auth/hooks/useLoginResolver";
-import { AuthCardContainer, AuthCard } from "../authLayout";
-import { TryAgainAlert } from "@lib/components/Alerts/TryAgainAlert";
-
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
 import Divider from "@mui/material/Divider";
 import TextField from "@mui/material/TextField";
 import Typography from "@mui/material/Typography";
-import InputAdornment from "@mui/material/InputAdornment";
-import IconButton from "@mui/material/IconButton";
+import { Link } from "react-router-dom";
+import { GoogleIcon, FacebookIcon } from "@lib/components/Auth/CustomIcons";
+import { useRegisterMutation } from "@/features/auth/hooks/useRegisterMutation";
+import {
+  RegisterSchema,
+  useRegisterResolver,
+} from "@/features/auth/hooks/useRegisterResolver";
+import { SubmitHandler, useController, useForm } from "react-hook-form";
+import { ApolloError } from "@apollo/client";
 import { VisibilityOff, Visibility } from "@mui/icons-material";
-import CircularProgress from "@mui/material/CircularProgress";
+import { InputAdornment, IconButton, CircularProgress } from "@mui/material";
+import { AuthCardContainer, AuthCard } from "../authLayout";
+import { TryAgainAlert } from "@lib/components/Alerts/TryAgainAlert";
 
-type Inputs = LoginSchema & {
+type Inputs = RegisterSchema & {
   onSubmitError?: string;
 };
 
-export const SignInCard = () => {
-  const [loginMutation, { loading }] = useLoginMutation();
+export const SignUpCard = (props: { disableCustomTheme?: boolean }) => {
   const [showPassword, setShowPassword] = useState(false);
-  const [isDilogOpen, setIsDilogOpen] = useState(false);
-  const [resolver] = useLoginResolver();
+  const [registerMutation, { loading }] = useRegisterMutation();
+  const [resolver] = useRegisterResolver();
 
   const {
     reset,
@@ -58,14 +53,16 @@ export const SignInCard = () => {
     control,
     defaultValue: "",
   });
-
-  const openChangePasswordDialog = () => {
-    setIsDilogOpen(true);
-  };
-
-  const closeChangePasswordDialog = () => {
-    setIsDilogOpen(false);
-  };
+  const { field: firstNameField, fieldState: firstNameState } = useController({
+    name: "firstName",
+    control,
+    defaultValue: "",
+  });
+  const { field: lastNameField, fieldState: lastNameState } = useController({
+    name: "lastName",
+    control,
+    defaultValue: "",
+  });
 
   const handleClickShowPassword = () => setShowPassword((show) => !show);
   const handleMouseDownPassword = (event: MouseEvent<HTMLButtonElement>) => {
@@ -80,7 +77,7 @@ export const SignInCard = () => {
     await trigger();
 
     try {
-      await loginMutation(formData);
+      await registerMutation(formData);
       // throw new Error("try try try");
     } catch (err) {
       reset();
@@ -99,11 +96,11 @@ export const SignInCard = () => {
     <AuthCardContainer direction="column" justifyContent="space-between">
       <AuthCard variant="outlined">
         <Typography
-          data-testid="login-title"
+          data-testid="register-title"
           component="h1"
           variant="h4"
           sx={{ width: "100%", fontSize: "clamp(2rem, 10vw, 2.15rem)" }}>
-          Sign in
+          Sign up
         </Typography>
         {onSubmitError && <TryAgainAlert message={onSubmitError.message} />}
         <form onSubmit={handleSubmit(onSubmit)}>
@@ -111,7 +108,7 @@ export const SignInCard = () => {
             required
             slotProps={{
               htmlInput: {
-                "data-testid": "login-username",
+                "data-testid": "register-username",
               },
             }}
             inputRef={usernameField.ref}
@@ -149,7 +146,7 @@ export const SignInCard = () => {
               type={showPassword ? "text" : "password"}
               slotProps={{
                 htmlInput: {
-                  "data-testid": "login-password",
+                  "data-testid": "register-password",
                 },
                 input: {
                   endAdornment: (
@@ -171,56 +168,93 @@ export const SignInCard = () => {
                 },
               }}
             />
-            <Button
-              component="button"
-              type="button"
-              onClick={openChangePasswordDialog}
-              sx={{ alignSelf: "baseline" }}>
-              Forgot your password?
-            </Button>
+          </Box>
+          <Box sx={{ margin: "14px auto" }}>
+            <TextField
+              required
+              slotProps={{
+                htmlInput: {
+                  "data-testid": "register-firstname",
+                },
+              }}
+              inputRef={firstNameField.ref}
+              name={firstNameField.name}
+              value={firstNameField.value}
+              onBlur={firstNameField.onBlur}
+              onChange={firstNameField.onChange}
+              label="First name"
+              placeholder="Enter your first name"
+              error={!!firstNameState.error}
+              helperText={
+                firstNameState.error ? firstNameState.error.message : null
+              }
+              color={firstNameState.error ? "error" : "primary"}
+              fullWidth
+              variant="outlined"
+            />
+          </Box>
+          <Box sx={{ margin: "14px auto" }}>
+            <TextField
+              required
+              slotProps={{
+                htmlInput: {
+                  "data-testid": "register-lastname",
+                },
+              }}
+              inputRef={lastNameField.ref}
+              name={lastNameField.name}
+              value={lastNameField.value}
+              onBlur={lastNameField.onBlur}
+              onChange={lastNameField.onChange}
+              label="Last name"
+              placeholder="Enter your last name"
+              error={!!lastNameState.error}
+              helperText={
+                lastNameState.error ? lastNameState.error.message : null
+              }
+              color={lastNameState.error ? "error" : "primary"}
+              fullWidth
+              variant="outlined"
+            />
           </Box>
           {/* <FormControlLabel
-          control={<Checkbox value="remember" color="primary" />}
-          label="Remember me"
-        /> */}
-          <ForgotPassword
-            open={isDilogOpen}
-            handleClose={closeChangePasswordDialog}
-          />
+            control={<Checkbox value="allowExtraEmails" color="primary" />}
+            label="I want to receive updates via email."
+          /> */}
           <Button
-            data-testid="login-submit"
+            data-testid="register-submit"
             type="submit"
             disabled={!isValid || loading}
             fullWidth
             variant="contained"
             endIcon={loading && <CircularProgress size={20} />}>
-            Sign in
+            Sign up
           </Button>
-          <Typography sx={{ textAlign: "center" }}>
-            Don&apos;t have an account?{" "}
-            <span>
-              <Link to="/register" style={{ alignSelf: "center" }}>
-                Sign up
-              </Link>
-            </span>
-          </Typography>
         </form>
-        <Divider>or</Divider>
+        <Divider>
+          <Typography sx={{ color: "text.secondary" }}>or</Typography>
+        </Divider>
         <Box sx={{ display: "flex", flexDirection: "column", gap: 2 }}>
           <Button
             fullWidth
             variant="outlined"
-            onClick={() => alert("Sign in with Google")}
+            onClick={() => alert("Sign up with Google")}
             startIcon={<GoogleIcon />}>
-            Sign in with Google
+            Sign up with Google
           </Button>
           <Button
             fullWidth
             variant="outlined"
-            onClick={() => alert("Sign in with Facebook")}
+            onClick={() => alert("Sign up with Facebook")}
             startIcon={<FacebookIcon />}>
-            Sign in with Facebook
+            Sign up with Facebook
           </Button>
+          <Typography sx={{ textAlign: "center" }}>
+            Already have an account?{" "}
+            <Link to="/login" style={{ alignSelf: "center" }}>
+              Sign in
+            </Link>
+          </Typography>
         </Box>
       </AuthCard>
     </AuthCardContainer>
