+++ client/src/features/auth/hooks/useRegisterResolver.tsx
