+++ client/src/features/createHuntDialog/stories/CreateHuntDialog.stories.tsx
@@ -1,9 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import React, { useState } from "react";
-import { userEvent, within, expect } from "@storybook/test";
 import Button from "@mui/material/Button";
 import { CreateHuntDialog as Component } from "../components/CreateHuntDialog";
-import dayjs from "dayjs";
 
 const meta: Meta = {
   title: "Components/Create Hunt Dailog",
