+++ client/src/features/dashboard/components/SideMenu/OptionsMenu.tsx
@@ -10,7 +10,7 @@ import ListItemIcon, { listItemIconClasses } from "@mui/material/ListItemIcon";
 import LogoutRoundedIcon from "@mui/icons-material/LogoutRounded";
 import MoreVertRoundedIcon from "@mui/icons-material/MoreVertRounded";
 import { MenuButton } from "../MenuButton";
-import { useLogoutMutation } from "@pages/dashboard/hooks/useLogoutMutation";
+import { useLogoutMutation } from "@/features/dashboard/hooks/useLogoutMutation";
 
 const MenuItem = styled(MuiMenuItem)({
   margin: "2px 0",
@@ -80,7 +80,6 @@ export const OptionsMenu = () => {
           <ListItemIcon>
             <LogoutRoundedIcon fontSize="small" />
           </ListItemIcon>
-
         </MenuItem>
       </Menu>
     </>
