+++ client/src/routes/AppRouter.tsx
@@ -1,19 +1,19 @@
 import React from "react";
 import "../../index.css";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
-import { Dashboard } from "@pages/dashboard/components/Dashboard";
+import { Dashboard } from "@/features/dashboard/components/Dashboard";
 import { catchallRouteToLogin } from "./catchAllRoutes/catchallRouteToLogin";
 import { catchallRouteToParent } from "./catchAllRoutes/catchallRouteToParent";
-import { AuthOutlet } from "@pages/auth/components/AuthOutlet";
-import { HuntPage } from "@/pages/hunts/components/HuntPage";
-import { HuntDetails } from "@/pages/hunts/components/HuntDetails";
-import { TeamsPage } from "@pages/teams/TeamsPage";
-import { CluesPage } from "@pages/clues/CluesPage";
-import { ResponsesPage } from "@pages/responses/ResponsesPage";
-import { SignInCard } from "@pages/auth/components/login/SignInCard";
-import { SignUpCard } from "@pages/auth/components/register/SignUpCard";
+import { AuthOutlet } from "@/features/auth/components/AuthOutlet";
+import { HuntPage } from "@/features/hunts/components/HuntPage";
+import { HuntDetails } from "@/features/hunts/components/HuntDetails";
+import { TeamsPage } from "@/features/teams/TeamsPage";
+import { CluesPage } from "@/features/clues/CluesPage";
+import { ResponsesPage } from "@/features/responses/ResponsesPage";
+import { SignInCard } from "@/features/auth/components/login/SignInCard";
+import { SignUpCard } from "@/features/auth/components/register/SignUpCard";
 import { DarkThemeProvider } from "@lib/context/DarkThemeProvider";
-import { HomePage } from "@pages/home/components/HomePage";
+import { HomePage } from "@/features/home/components/HomePage";
 
 const router = createBrowserRouter([
   {
