+++ client/tsconfig.json
@@ -15,8 +15,8 @@
       "@generated/*": ["src/generated/*"],
       "@graphql/*": ["src/graphql/*"],
       "@lib/*": ["src/lib/*"],
-      "@pages/*": ["src/pages/*"],
-      "@routes/*": ["src/routes/*"]
+      "@routes/*": ["src/routes/*"],
+      "@features/*": ["src/features/*"]
     }
   },
   "exclude": ["node_modules", "src/dux"]
